requirements.txt
anyio==4.3.0
argostranslate==1.9.3
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ctranslate2==3.20.0
Django==5.0.4
django-debug-toolbar==4.3.0
dnspython==2.6.1
filelock==3.13.4
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
pymongo==4.7.0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.4
regex==2024.4.16
requests==2.31.0
sacremoses==0.0.53
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
stanza==1.1.1
sympy==1.12
tbb==2021.12.0
torch==2.3.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
whitenoise==6.6.0
